move some tests to __internal__
