+++ test/__internal__/child_process_utils/child_process_import_namespace/child_process_import_namespace.test.js
@@ -1,0 +1,10 @@
+import { assert } from "@jsenv/assert"
+import { resolveUrl } from "@jsenv/filesystem"
+
+import { importUsingChildProcess } from "@jsenv/core"
+
+const fileUrl = resolveUrl("./file.js", import.meta.url)
+
+const actual = await importUsingChildProcess(fileUrl)
+const expected = { toto: "yep" }
+assert({ actual, expected })
