+++ test/__internal__/child_process_utils/child_process_import_production/root/package.json
