+++ test/__internal__/child_process_utils/child_process_require_namespace/child_process_require_namespace.test.js
@@ -1,0 +1,10 @@
+import { assert } from "@jsenv/assert"
+import { resolveUrl } from "@jsenv/filesystem"
+
+import { requireUsingChildProcess } from "@jsenv/core"
+
+const fileUrl = resolveUrl("./file.cjs", import.meta.url)
+
+const actual = await requireUsingChildProcess(fileUrl)
+const expected = { toto: "yep" }
+assert({ actual, expected })
