+++ test/__internal__/child_process_utils/child_process_require_namespace/file.cjs
