+++ test/__internal__/child_process_utils/child_process_require_production/child_process_require_production.test.js
@@ -1,0 +1,12 @@
+import { assert } from "@jsenv/assert"
+import { resolveUrl } from "@jsenv/filesystem"
+
+import { requireUsingChildProcess } from "@jsenv/core"
+
+const fileUrl = resolveUrl("./root/main.cjs", import.meta.url)
+
+const actual = await requireUsingChildProcess(fileUrl, {
+  commandLineOptions: ["--conditions=production"],
+})
+const expected = { env: "prod" }
+assert({ actual, expected })
