+++ test/__internal__/child_process_utils/child_process_require_production/root/package.json
