--- test/child-process-import/child-process-import-namespace/child-process-import-namespace.test.js
@@ -1,11 +1,0 @@
-import { importUsingChildProcess } from "@jsenv/core"
-import { assert } from "@jsenv/assert"
-import { resolveUrl } from "@jsenv/filesystem"
-
-const fileUrl = resolveUrl("./file.js", import.meta.url)
-
-{
-  const actual = await importUsingChildProcess(fileUrl)
-  const expected = { toto: "yep" }
-  assert({ actual, expected })
-}
