--- test/child-process-import/child-process-import-production/child-process-import-production.test.js
@@ -1,13 +1,0 @@
-import { importUsingChildProcess } from "@jsenv/core"
-import { assert } from "@jsenv/assert"
-import { resolveUrl } from "@jsenv/filesystem"
-
-const fileUrl = resolveUrl("./root/main.js", import.meta.url)
-
-{
-  const actual = await importUsingChildProcess(fileUrl, {
-    commandLineOptions: ["--conditions=production"],
-  })
-  const expected = { env: "prod" }
-  assert({ actual, expected })
-}
