--- test/child-process-require/child-process-require-namespace/child-process-require-namespace.test.js
@@ -1,11 +1,0 @@
-import { requireUsingChildProcess } from "@jsenv/core"
-import { assert } from "@jsenv/assert"
-import { resolveUrl } from "@jsenv/filesystem"
-
-const fileUrl = resolveUrl("./file.cjs", import.meta.url)
-
-{
-  const actual = await requireUsingChildProcess(fileUrl)
-  const expected = { toto: "yep" }
-  assert({ actual, expected })
-}
