--- test/child-process-require/child-process-require-production/child-process-require-production.test.js
@@ -1,13 +1,0 @@
-import { requireUsingChildProcess } from "@jsenv/core"
-import { assert } from "@jsenv/assert"
-import { resolveUrl } from "@jsenv/filesystem"
-
-const fileUrl = resolveUrl("./root/main.cjs", import.meta.url)
-
-{
-  const actual = await requireUsingChildProcess(fileUrl, {
-    commandLineOptions: ["--conditions=production"],
-  })
-  const expected = { env: "prod" }
-  assert({ actual, expected })
-}
